refactor: info 기능 패키지 구조 리팩토링
