--- application/app/src/main/java/com/example/priend/main/info/recyclerview/InfoItemType.kt
@@ -1,6 +1,0 @@
-package com.example.priend.main.info.recyclerview
-
-enum class InfoItemType {
-    ADD,
-    PLANT
-}
