+++ application/app/src/main/java/com/example/priend/main/info/view/recyclerview/InfoItem.kt
@@ -1,4 +1,4 @@
-package com.example.priend.main.info.recyclerview
+package com.example.priend.main.info.view.recyclerview
 
 data class InfoItem(
     val type : InfoItemType,
