+++ application/app/src/main/java/com/example/priend/main/info/view/recyclerview/InfoItemType.kt
@@ -1,0 +1,6 @@
+package com.example.priend.main.info.view.recyclerview
+
+enum class InfoItemType {
+    ADD,
+    PLANT
+}
