+++ application/app/src/main/java/com/example/priend/main/info/view/recyclerview/viewholder/InfoPlantViewHolder.kt
@@ -1,10 +1,10 @@
-package com.example.priend.main.info.recyclerview.viewholder
+package com.example.priend.main.info.view.recyclerview.viewholder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.example.priend.common.view.RecyclerViewItemClickListener
 import com.example.priend.databinding.ItemInfoBinding
-import com.example.priend.main.info.recyclerview.InfoItem
+import com.example.priend.main.info.view.recyclerview.InfoItem
 
 class InfoPlantViewHolder(
     private val binding: ItemInfoBinding,
