+++ application/app/src/main/res/layout/fragment_info.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".main.info.InfoFragment">
+    tools:context=".main.info.view.InfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
