+++ application/app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@id/infoFragment"
-        android:name="com.example.priend.main.info.InfoFragment"
+        android:name="com.example.priend.main.info.view.InfoFragment"
         android:label="fragment_info"
         tools:layout="@layout/fragment_info"
         />
